Trait genio::bufio::BufReadProgress

pub trait BufReadProgress: BufRead {
    type BufReadError;

    fn fill_progress(
        &mut self
    ) -> Result<&[u8], BufError<Self::BufReadError, Self::ReadError>>;
}

When reading from reader, sometimes it’s beneficial to read n bytes at once. However, BufRead itself doesn’t guarantee that more bytes will be available when calling fill_buf multiple times. This trait provides fill_progress function with that guarantee.

Required Associated Types

type BufReadError

Error that occurs in buffer itself. Most often if buffer is out of memory.

Required Methods

fn fill_progress(
&mut self
) -> Result<&[u8], BufError<Self::BufReadError, Self::ReadError>>

Fills the internal buffer guaranteeing that successive calls to this function return more and more bytes (or an error).