Trait genio::bufio::BufWriteRequire

pub unsafe trait BufWriteRequire: BufWrite {
    type BufWriteError;

    fn require_buffer(
        &mut self,
        size: usize
    ) -> Result<*mut [u8], Self::BufWriteError>;
}

This trait allows requiring buffer of specified size.

Required Associated Types

type BufWriteError

Indicates error in buffer. Most often caused by size being too large but it might be also a write error, if buf writer tried to flush buffer.

Required Methods

fn require_buffer(
&mut self,
size: usize
) -> Result<*mut [u8], Self::BufWriteError>

Require buffer with minimum size bytes. It is an error to return smaller buffer but unsafe code can’t rely on it.